Compiler refactored.

+++ interpreter/test/test_char_operand.cpp
@@ -404,7 +404,7 @@ test_array_read_value(DArray array, const DBS_T testVal)
 
   for (int index = array.Count() - 1; index >= 0; --index)
     {
-      if (iterator.Operand().IteratorOffset() != index)
+      if ((int)iterator.Operand().IteratorOffset() != index)
         return false;
 
       DBS_T first, second;
@@ -564,7 +564,7 @@ test_table_value(TableOperand&  tableOp,
 
   for (int i = array.Count() - 1; i >= 0; --i)
     {
-      if (temp.Operand().IteratorOffset() != i)
+      if ((int)temp.Operand().IteratorOffset() != i)
         return false;
 
       DBS_T ref, val;
